firestore_python_code/requirements.txt
google-cloud-firestore==2.6.1